requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.1
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
pyairtable==1.2.0
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
yarl==1.8.1
